+++ Source/Data/Backend/index.ts
@@ -12,9 +12,6 @@ const swaggerDoc = require('./swagger.json');
     await Host.start({
         swaggerDoc,
         graphQLResolvers: queries,
-        dolittleCallback: (client) => {
-            console.log('Hello');
-        },
         expressCallback: (app) => {
             RegisterRoutes(app);
         }
