+++ Source/Data/Backend/queries.ts
@@ -1,10 +1,9 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import { ApplicationForListingQueries } from './listing/ApplicationForListingQueries';
+import { ApplicationForListingQueries } from './backups/ApplicationForListingQueries';
 import { BackupForListingQueries } from './backups/BackupForListingQueries';
-import { TenantQueries } from './backups/TenantQueries';
 
 export default [
-    ApplicationForListingQueries, BackupForListingQueries, TenantQueries
+    ApplicationForListingQueries, BackupForListingQueries
 ];
